improve type checking and resolve errors
