+++ web/src/lib/components/admin-page/settings/storage-template/supported-datetime-panel.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { SystemConfigTemplateStorageOptionDto } from '@api';
+	import type { SystemConfigTemplateStorageOptionDto } from '@api';
 	import * as luxon from 'luxon';
 
 	export let options: SystemConfigTemplateStorageOptionDto;
