+++ web/src/lib/components/album-page/album-card.ts
@@ -1,0 +1,12 @@
+import type { AlbumResponseDto } from '@api';
+
+export type OnShowContextMenu = {
+	showalbumcontextmenu: OnShowContextMenuDetail;
+};
+
+export type OnClick = {
+	click: OnClickDetail;
+};
+
+export type OnShowContextMenuDetail = { x: number; y: number };
+export type OnClickDetail = AlbumResponseDto;
