+++ web/src/lib/models/asset-grid-state.ts
@@ -1,4 +1,4 @@
-import { AssetResponseDto } from '@api';
+import type { AssetResponseDto } from '@api';
 
 export class AssetBucket {
 	/**
