+++ web/src/lib/stores/archived-asset.store.ts
@@ -1,4 +1,4 @@
-import { AssetResponseDto } from '@api';
+import type { AssetResponseDto } from '@api';
 import { writable } from 'svelte/store';
 
 export const archivedAsset = writable<AssetResponseDto[]>([]);
