+++ web/src/lib/stores/preferences.store.ts
@@ -1,5 +1,4 @@
 import { browser } from '$app/environment';
-import { MapSettings } from '$lib/components/map-page/map-settings-modal.svelte';
 import { persisted } from 'svelte-local-storage-store';
 
 const initialTheme =
@@ -20,6 +19,14 @@ export const locale = persisted<string | undefined>('locale', undefined, {
 		stringify: (obj) => obj ?? ''
 	}
 });
+
+export interface MapSettings {
+	allowDarkMode: boolean;
+	onlyFavorites: boolean;
+	relativeDate: string;
+	dateAfter: string;
+	dateBefore: string;
+}
 
 export const mapSettings = persisted<MapSettings>('map-settings', {
 	allowDarkMode: true,
