+++ web/src/routes/(user)/albums/albums.bloc.ts
@@ -1,10 +1,10 @@
+import type { OnShowContextMenuDetail } from '$lib/components/album-page/album-card';
 import {
 	notificationController,
 	NotificationType
 } from '$lib/components/shared-components/notification/notification';
 import { AlbumResponseDto, api } from '@api';
-import { OnShowContextMenuDetail } from '$lib/components/album-page/album-card.svelte';
-import { writable, derived, get } from 'svelte/store';
+import { derived, get, writable } from 'svelte/store';
 
 type AlbumsProps = { albums: AlbumResponseDto[] };
