+++ web/src/test-data/factories/album-factory.ts
@@ -1,6 +1,7 @@
-import { AlbumResponseDto } from '@api';
-import { Sync } from 'factory.ts';
+import type { AlbumResponseDto } from '@api';
 import { faker } from '@faker-js/faker';
+import { Sync } from 'factory.ts';
+import { userFactory } from './user-factory';
 
 export const albumFactory = Sync.makeFactory<AlbumResponseDto>({
 	albumName: Sync.each(() => faker.commerce.product()),
@@ -8,8 +9,10 @@ export const albumFactory = Sync.makeFactory<AlbumResponseDto>({
 	assetCount: Sync.each((i) => i % 5),
 	assets: [],
 	createdAt: Sync.each(() => faker.date.past().toISOString()),
+	updatedAt: Sync.each(() => faker.date.past().toISOString()),
 	id: Sync.each(() => faker.datatype.uuid()),
 	ownerId: Sync.each(() => faker.datatype.uuid()),
+	owner: userFactory.build(),
 	shared: false,
 	sharedUsers: []
 });
